refactor: inject dependencies into SignInViewModel
This commit refactors `SignInViewModel` to use dependency injection for `FirebaseAuth` and `GoogleIdOptionFactory`.

This change allows for easier testing by enabling the injection of mocks, removing the need for `mockkStatic` in the tests. A `GoogleIdOptionFactory` interface and its default implementation have been introduced to abstract the creation of `GetGoogleIdOption`. The corresponding unit tests have been updated to reflect these changes.